+++ build-logic/convention/src/main/kotlin/com/rokt/roktux/AndroidCompose.kt
@@ -8,13 +8,11 @@ import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 import java.io.File
 
 fun Project.configureAndroidCompose(commonExtension: CommonExtension<*, *, *, *, *, *>) {
+    pluginManager.apply("org.jetbrains.kotlin.plugin.compose")
+
     commonExtension.apply {
         buildFeatures {
             compose = true
-        }
-
-        composeOptions {
-            kotlinCompilerExtensionVersion = libs.findVersion("androidxComposeCompiler").get().toString()
         }
 
         dependencies {