+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update Jetpack Compose BOM to 2026.05.01.
+
 ## [0.9.3] - 2026-05-13
 
 ### Fixed